requirements.txt
aioredis==1.2.0
asgiref==3.1.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.6.2
Automat==0.7.0
beautifulsoup4==4.7.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
Cython==0.29.21
daphne==2.4.1
Django==2.2.2
django-ckeditor==5.9.0
django-crispy-forms==1.10.0
django-filter==2.2.0
django-filters==0.2.1
django-js-asset==1.2.2
djangorestframework==3.11.0
docopt==0.6.2
et-xmlfile==1.0.1
google-cloud==0.34.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4.1
Js2Py==0.66
msgpack==0.6.2
openpyxl==3.0.5
packaging==19.0
pdfminer.six==20201018
Pillow==6.2.1
pipwin==0.4.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.6.1
PyHamcrest==1.9.0
pyjsparser==2.7.1
PyMuPDF==1.18.4
pyOpenSSL==19.1.0
pyparsing==2.4.0
PyPDF2==1.26.0
PyPrind==2.11.2
pytesseract==0.3.6
pytz==2018.9
redis==3.4.1
requests==2.22.0
robobrowser==0.5.3
service-identity==18.1.0
six==1.12.0
sortedcontainers==2.3.0
soupsieve==1.9.2
SpeechRecognition==3.8.1
sqlparse==0.3.0
stripe==2.50.0
twisted==19.10.0
txaio==18.8.1
tzlocal==1.5.1
urllib3==1.25.3
virtualenv==16.4.3
virtualenvwrapper-win==1.2.5
Werkzeug==0.15.4
zope.interface==4.6.0
